+++ packages/react/src/components/PlaceholderCommon.tsx
@@ -155,9 +155,12 @@ export class PlaceholderCommon<T extends PlaceholderProps> extends React.Compone
 
   getSXAParams(rendering: ComponentRendering) {
     if (!rendering.params) return {};
+
+    const { GridParameters, Styles } = rendering.params;
+
     return (
-      rendering.params.FieldNames && {
-        styles: `${rendering.params.GridParameters || ''} ${rendering.params.Styles || ''}`,
+      (GridParameters || Styles) && {
+        styles: `${GridParameters || ''} ${Styles || ''}`,
       }
     );
   }
@@ -239,7 +242,6 @@ export class PlaceholderCommon<T extends PlaceholderProps> extends React.Compone
         );
 
         if (!isEmpty) {
-          console.log('test');
           // assign type based on passed element - type='text/sitecore' should be ignored when renderEach Placeholder prop function is being used
           const type = rendered.props.type === 'text/sitecore' ? rendered.props.type : '';
 
@@ -247,6 +249,11 @@ export class PlaceholderCommon<T extends PlaceholderProps> extends React.Compone
 
           // wrapping with error boundary could cause problems in case where parent component uses withPlaceholder HOC and tries to access its children props
           // that's why we need to expose element's props here
+          const isByocWrapper = componentRendering.componentName === BYOC_WRAPPER_RENDERING_NAME;
+
+          // all dynamic elements will have a separate render prop
+          const isDynamicComponent = !!(component as LazyComponentType).render?.preload;
+
           rendered = (
             <ErrorBoundary
               data-testid="error-boundary"
@@ -254,7 +261,7 @@ export class PlaceholderCommon<T extends PlaceholderProps> extends React.Compone
               errorComponent={this.props.errorComponent}
               componentLoadingMessage={this.props.componentLoadingMessage}
               type={type}
-              isDynamic={(component as LazyComponentType).render?.preload ? true : false}
+              isDynamic={isDynamicComponent || isByocWrapper}
               disableSuspense={disableSuspense}
               {...rendered.props}
             >